Add support for ActiveSupport notifications
In taskrabbit/makara#103 an attempt was made to implement a bunch of
DTrace probes to monitor makara. Unfortunately, DTrace isn't something
most Rails developers are familiar with so the adoption of this might be
difficult to get off the ground. Here I've swapped the DTrace probes to
use ActiveSupport notifications which you are able to subscribe to
(usually within an initializer) and can offload to the monitoring tool
of your choose without needing to touch the library itself. For example:

```
# config/initializers/makara_instrumentation.rb
ActiveSupport::Notifications.subscribe(/makara/) do |*args|
  # .. do something with args
end
```

Will allow you to capture all makara events and perform additional calls
or monitoring on anything you deem important. If you are only interested
in a subset of the events, you can do that too.

```
# config/initializers/makara_instrumentation.rb
ActiveSupport::Notifications.subscribe('makara.proxy') do |*args|
  # .. do something with args on 'makara.proxy' events
end
```