Convert Context into a stickiness state representation
Instead of keeping track of two MD5 hashes that allow us to be stuck to
master for 1 or 2 requests at most, keep whether we are currently stuck
or not, for each configuration.

The context represents the current stickiness state. It knows whether we
are stuck to master for each of the different Makara configurations we
have defined. It's persisted between requests via an only cookie we set,
whose max age depends on stickiness duration for each of the
configurations we store at any given moment.