Stop referring to the proxy id as "config id", it's very confusing
The id defined in the configuration for a proxy was being referred as
`config_id` in several places, so I had kept it, but it was certainly
confusing, it wasn't easy to know what I was talking about. That id
really identifies the proxy, so there is no point on using a misleading
name for it.